switched output block state to using ItemStack, refacto, removed useless class
